Move time filter into new component style, making space for switch logic in the higher level component.
